pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "truffle-autoresearch"
version = "0.1.0"
description = "Automated ML research on personal hardware, coordinated by a Truffle device"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "paramiko>=3.4.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
autoresearch = "truffle_autoresearch.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/truffle_autoresearch"]
